requirement.txt
accelerate==0.24.1
azure-core==1.29.5
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
diffuser==0.0.1
diffusers==0.23.1
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pika==1.3.2
Pillow==10.1.0
psutil==5.9.6
pycparser==2.21
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
six==1.16.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
urllib3==2.1.0
zipp==3.17.0
